server/requirements.txt
alembic==0.9.9
pylint==1.9.1
flask==1.0.2
Flask-Script==2.0.6
coverage==4.5.1
sqlalchemy==1.2.8
blueprints==2.3.0.2
Flask-SQLAlchemy==2.3.2
flask-sqlacodegen==1.1.6.1
flask-bootstrap==3.3.7.1
flask-sse==0.2.1
Flask-Migrate==2.1.1
uwsgi==2.0.17
gevent==1.4.0
psycopg2==2.6.1
Werkzeug==0.14.1
flask-cors==2.1.0
amqp==2.2.2
Babel==2.5.3
billiard==3.5.0.3
celery==4.2.1
flower==0.9.2
kombu==4.2.0
pytz==2018.3
redis==2.10.6
tornado==5.0.2
vine==1.1.4
python-dotenv==0.10.1
flask-socketio
moviepy==0.2.3.5
pafy==0.5.4
youtube-dl==2018.12.17